[CoreBundle] Moving some workspace API code.
