+++ main/core/Resources/config/routing.yml
@@ -306,14 +306,6 @@ claro_facet_api:
     options:
         expose: true
 
-claro_workspace_api:
-    resource: "@ClarolineCoreBundle/Controller/API/Admin/WorkspaceController.php"
-    type: rest
-    prefix: api
-    defaults: {_format: json}
-    options:
-        expose: true
-
 claro_group_api:
     resource: "@ClarolineCoreBundle/Controller/API/User/GroupController.php"
     type: rest
